Everything is better now :)

+++ src/activations.cu
@@ -10,10 +10,10 @@
  * sigmoidFunction - a Sigmoid activation function
  * __host__ decoration tells NVCC this function should run on CPU, and be callable from the CPU host
  * __device__ decoration tells NVCC this function should run on GPU, and be callable from the GPU device
- * @params: d - a double value
+ * @params: d - a float value
  * @returns: the Sigmoid of d
  */
-__host__ __device__ double sigmoidFunction(double d) {
+__host__ __device__ float sigmoidFunction(float d) {
     return 1.0 / (1.0 + exp(-1.0 * d));
 } //end sigmoid activation function
 
@@ -21,10 +21,10 @@ __host__ __device__ double sigmoidFunction(double d) {
  * sigmoidDerivative - a Sigmoid derivative function
  * __host__ decoration tells NVCC this function should run on CPU, and be callable from the CPU host
  * __device__ decoration tells NVCC this function should run on GPU, and be callable from the GPU device
- * @params: d - a double value
+ * @params: d - a float value
  * @returns: the Sigmoid derivative of d
  */
-__host__ __device__ double sigmoidDerivative(double d) {
+__host__ __device__ float sigmoidDerivative(float d) {
     return sigmoidFunction(d) * (1.0 - sigmoidFunction(d));
 } //end sigmoid derivative function
 
@@ -32,10 +32,10 @@ __host__ __device__ double sigmoidDerivative(double d) {
  * tanhFunction - a TanH activation function
  * __host__ decoration tells NVCC this function should run on CPU, and be callable from the CPU host
  * __device__ decoration tells NVCC this function should run on GPU, and be callable from the GPU device
- * @params: d - a double value
+ * @params: d - a float value
  * @returns: the TanH of d
  */
-__host__ __device__ double tanhFunction(double d) {
+__host__ __device__ float tanhFunction(float d) {
     return (2.0 / (1.0 + exp(-2.0 * d))) - 1.0;
 } //end tanh activation function
 
@@ -43,10 +43,10 @@ __host__ __device__ double tanhFunction(double d) {
  * tanhDerivative - a TanH derivative function
  * __host__ decoration tells NVCC this function should run on CPU, and be callable from the CPU host
  * __device__ decoration tells NVCC this function should run on GPU, and be callable from the GPU device
- * @params: d - a double value
+ * @params: d - a float value
  * @returns: the TanH derivative of d
  */
-__host__ __device__ double tanhDerivative(double d) {
+__host__ __device__ float tanhDerivative(float d) {
     return 1.0 - pow(tanhFunction(d), 2.0);
 } //end tanh derivative function
 
@@ -54,10 +54,10 @@ __host__ __device__ double tanhDerivative(double d) {
  * reluFunction - a ReLU activation function
  * __host__ decoration tells NVCC this function should run on CPU, and be callable from the CPU host
  * __device__ decoration tells NVCC this function should run on GPU, and be callable from the GPU device
- * @params: d - a double value
+ * @params: d - a float value
  * @returns: the ReLU of d
  */
-__host__ __device__ double reluFunction(double d) {
+__host__ __device__ float reluFunction(float d) {
     if (d < 0) {
         return 0.0;
     } else {
@@ -69,10 +69,10 @@ __host__ __device__ double reluFunction(double d) {
  * reluDerivative - a ReLU derivative function
  * __host__ decoration tells NVCC this function should run on CPU, and be callable from the CPU host
  * __device__ decoration tells NVCC this function should run on GPU, and be callable from the GPU device
- * @params: d - a double value
+ * @params: d - a float value
  * @returns: the ReLU derivative of d
  */
-__host__ __device__ double reluDerivative(double d) {
+__host__ __device__ float reluDerivative(float d) {
     if (d < 0) {
         return 0.0;
     } else {
@@ -83,11 +83,11 @@ __host__ __device__ double reluDerivative(double d) {
 /*
  * cudaKernel_ActivateLayerUsingSigmoid - applies Sigmoid activation to every neuron in a layer
  * __global__ decoration tells NVCC this function should run on GPU, and be callable from the CPU host
- * @params: devNeurons - a pointer to an array of double values in GPU device memory
+ * @params: devNeurons - a pointer to an array of float values in GPU device memory
  * @params: indexOfFirstNeuronInLayer - the index of the first neuron in the layer
  * @params: numberOfNeuronsInLayer - the total number of neurons in the layer
  */
-__global__ void cudaKernel_ActivateLayerUsingSigmoid(double* devNeurons, int indexOfFirstNeuronInLayer, int numberOfNeuronsInLayer) {
+__global__ void cudaKernel_ActivateLayerUsingSigmoid(float* devNeurons, unsigned int indexOfFirstNeuronInLayer, unsigned int numberOfNeuronsInLayer) {
     volatile unsigned int id = threadIdx.x + blockIdx.x * blockDim.x;
 
     // check that this thread is within our desired range (extra threads may have been launched for GPU optimization)
@@ -99,11 +99,11 @@ __global__ void cudaKernel_ActivateLayerUsingSigmoid(double* devNeurons, int ind
 /*
  * cudaKernel_ActivateLayerUsingTanh - applies TanH activation to every neuron in a layer
  * __global__ decoration tells NVCC this function should run on GPU, and be callable from the CPU host
- * @params: devNeurons - a pointer to an array of double values in GPU device memory
+ * @params: devNeurons - a pointer to an array of float values in GPU device memory
  * @params: indexOfFirstNeuronInLayer - the index of the first neuron in the layer
  * @params: numberOfNeuronsInLayer - the total number of neurons in the layer
  */
-__global__ void cudaKernel_ActivateLayerUsingTanh(double* devNeurons, int indexOfFirstNeuronInLayer, int numberOfNeuronsInLayer) {
+__global__ void cudaKernel_ActivateLayerUsingTanh(float* devNeurons, unsigned int indexOfFirstNeuronInLayer, unsigned int numberOfNeuronsInLayer) {
     volatile unsigned int id = threadIdx.x + blockIdx.x * blockDim.x;
 
     // check that this thread is within our desired range (extra threads may have been launched for GPU optimization)
@@ -115,11 +115,11 @@ __global__ void cudaKernel_ActivateLayerUsingTanh(double* devNeurons, int indexO
 /*
  * cudaKernel_ActivateLayerUsingRelu - applies ReLU activation to every neuron in a layer
  * __global__ decoration tells NVCC this function should run on GPU, and be callable from the CPU host
- * @params: devNeurons - a pointer to an array of double values in GPU device memory
+ * @params: devNeurons - a pointer to an array of float values in GPU device memory
  * @params: indexOfFirstNeuronInLayer - the index of the first neuron in the layer
  * @params: numberOfNeuronsInLayer - the total number of neurons in the layer
  */
-__global__ void cudaKernel_ActivateLayerUsingRelu(double* devNeurons, int indexOfFirstNeuronInLayer, int numberOfNeuronsInLayer) {
+__global__ void cudaKernel_ActivateLayerUsingRelu(float* devNeurons, unsigned int indexOfFirstNeuronInLayer, unsigned int numberOfNeuronsInLayer) {
     volatile unsigned int id = threadIdx.x + blockIdx.x * blockDim.x;
 
     // check that this thread is within our desired range (extra threads may have been launched for GPU optimization)
